backend/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
bcrypt==4.1.1
requests==2.31.0
